% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvgb.R
\docType{package}
\name{mvgb}
\alias{mvgb}
\title{Multivariate Probabilities of Scale Mixtures of Multivariate Normal Distributions}
\description{
The QRVSN algorithm is used in a broader context.
}
\section{The mvgb functions include}{


\code{\link{pmvt}} -- multivariate student t distribution probabilities

\code{\link{rmvt}} -- multivariate student t distribution random variates
}

