% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvgb.R
\docType{package}
\name{mvgb}
\alias{mvgb}
\title{Multivariate Probabilities of Scale Mixtures of Multivariate Normal Distributions}
\description{
The QRVSN algorithm is used in a broader context: using the Genz and Bretz
(2002) algorithm to calculate multivariate distribution probabilities.
}
\section{Multivariate Subgaussian Stable Distribution}{


\code{\link{pmvss}} -- multivariate subgaussian stable distribution probabilities
}

