% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvgb.R
\docType{package}
\name{mvgb}
\alias{mvgb}
\title{mvgb: A package for computating the notorious bar statistic.}
\description{
The mvgb package provides three categories of important functions:
foo, bar and baz.
}
\section{Mypackage functions}{

The mvgb functions ...
}

