% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvgb.R
\docType{package}
\name{mvgb}
\alias{mvgb-package}
\title{Multivariate Probabilities of Scale Mixtures of Multivariate Normal Distributions}
\description{
The QRVSN algorithm is used in a broader context: using the Genz and Bretz
(2002) algorithm to calculate multivariate distribution probabilities.
}
\section{Multivariate Subgaussian Stable Distribution}{


\code{\link{pmvss}} -- multivariate subgaussian stable distribution probabilities
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/swihart/mvgb}
  \item Report bugs at \url{https://github.com/swihart/mvgb/issues}
}

}
\author{
\strong{Maintainer}: Bruce Swihart \email{bruce.swihart@gmail.com} (\href{https://orcid.org/0000-0002-4216-9942}{ORCID})

Authors:
\itemize{
  \item Alan Genz (wrote mvtdstpack.f)
}

}
